ai.hohyub.site/diffusers/requirements.txt
# PyTorch 2.5.1 + CUDA 12.1 (RTX 4060 최적화, SDXL 안정적)
# 설치: pip install torch==2.5.1+cu121 torchvision==0.20.1+cu121 torchaudio==2.5.1+cu121 --index-url https://download.pytorch.org/whl/cu121
# conda diffuser 환경의 공통 라이브러리
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1
diffusers>=0.36.0
transformers>=4.57.3
accelerate>=1.12.0
safetensors>=0.7.0
pillow>=12.0.0
numpy>=2.2.6
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.21
python-dotenv>=1.2.1
huggingface-hub>=0.36.0
pydantic>=2.12.5
boto3>=1.34.0